% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{AddHealth}
\alias{AddHealth}
\title{Adolescent to Adult Health (Wave I)}
\format{A data frame with 6504 observations on 2829 variables}
\source{
\url{http://www.cpc.unc.edu/projects/addhealth}
}
\usage{
AddHealth
}
\description{
The National Longitudinal Study of Adolescent to Adult Health (Add Health) is a longitudinal study of a nationally representative sample of adolescents in grades 7-12 in the United States during the 1994-95 school year.
}
\details{
See the code books in inst/CodeBooks for details
}
\examples{
dim(AddHealth)
}
\keyword{datasets}

