% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{addhealth_public4}
\alias{addhealth_public4}
\title{Adolescent to Adult Health (Wave 4)}
\format{A data frame with 6504 observations on 976 variables}
\source{
\url{http://www.cpc.unc.edu/projects/addhealth}
}
\usage{
addhealth_public4
}
\description{
The National Longitudinal Study of Adolescent to Adult Health (Add Health) is a longitudinal study of a nationally representative sample of adolescents in grades 7-12 in the United States during the 1994-95 school year.
}
\details{
See the code books in the CodeBooks folder for details.
The CodeBooks folder is a top level folder of the \code{PDS} package.  To see where the \code{PDS} package is
installed on your machine, type the following at the \code{R} prompt: \code{system.file(package = "PDS")}
}
\examples{
dim(addhealth_public4)
}
\keyword{datasets}

