% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{frustration}
\alias{frustration}
\title{Frustation by Academic Major}
\format{\preformatted{'data.frame':	140 obs. of  6 variables:
 $ Business         : int  11 6 6 4 6 9 8 3 11 12 ...
 $ English          : int  11 9 14 13 9 12 10 12 9 11 ...
 $ Mathematics      : int  9 16 11 11 12 17 12 14 10 12 ...
 $ Psychology       : int  11 19 13 10 14 10 13 14 15 17 ...
 $ Frustration.Score: int  11 6 6 4 6 9 8 3 11 12 ...
 $ Major            : Factor w/ 4 levels "Business","English",..: 1 1 1 1 1 1 1 1 1 1 ...
}}
\source{
\url{https://oli.cmu.edu/jcourse/workbook/activity/page?context=434b8ad280020ca60166a71faa2caf60}
}
\usage{
frustration
}
\description{
Frustation by Academic Major
}
\examples{
boxplot(Frustration.Score ~ Major, data = frustration)
}
\keyword{datasets}

