% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Country level statistics}
\format{\preformatted{'data.frame':	213 obs. of  16 variables:
 $ country             : Factor w/ 213 levels "Afghanistan",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ incomeperperson     : num  NA 1915 2232 21943 1381 ...
 $ alcconsumption      : num  0.03 7.29 0.69 10.17 5.57 ...
 $ armedforcesrate     : num  0.57 1.02 2.31 NA 1.46 ...
 $ breastcancerper100th: num  26.8 57.4 23.5 NA 23.1 NA 73.9 51.6 NA 83.2 ...
 $ co2emissions        : num  7.59e+07 2.24e+08 2.93e+09 NA 2.48e+08 ...
 $ femaleemployrate    : num  25.6 42.1 31.7 NA 69.4 ...
 $ hivrate             : num  NA NA 0.1 NA 2 NA 0.5 0.1 NA 0.1 ...
 $ internetuserate     : num  3.65 44.99 12.5 81 10 ...
 $ lifeexpectancy      : num  48.7 76.9 73.1 NA 51.1 ...
 $ oilperperson        : num  NA NA 0.42 NA NA ...
 $ polityscore         : int  0 9 2 NA -2 NA 8 5 NA 10 ...
 $ relectricperperson  : num  NA 636 591 NA 173 ...
 $ suicideper100th     : num  6.68 7.7 4.85 5.36 14.55 ...
 $ employrate          : num  55.7 51.4 50.5 NA 75.7 ...
 $ urbanrate           : num  24 46.7 65.2 88.9 56.7 ...
}}
\source{
\url{http://www.gapminder.org}
}
\usage{
gapminder
}
\description{
Country level statistics
}
\details{
See the code books in inst/CodeBooks for details
}
\examples{
dim(gapminder)
}
\keyword{datasets}

