% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{ipeds_pds}
\alias{ipeds_pds}
\title{Integrated Post-secondary Education Data System}
\format{A data frame with 7764 observations on 544 variables}
\source{
\url{http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
ipeds_pds
}
\description{
Integrated Post-secondary Education Data System
}
\details{
IPEDS \url{http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
is the primary source for data on colleges, universities, and technical
and vocational postsecondary institutions in the United States.
See the code books in the CodeBooks folder for details.
The CodeBooks folder is a top level folder of the \code{PDS} package.  To see where the \code{PDS} package is
installed on your machine, type the following at the \code{R} prompt: \code{system.file(package = "PDS")}
}
\examples{
dim(ipeds_pds)
}
\keyword{datasets}

