% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{marscrater_pds}
\alias{marscrater_pds}
\title{Mars Craters}
\format{A data frame with 384343 observations on 10 variables}
\usage{
marscrater_pds
}
\description{
marscrater
}
\details{
See the code books in the CodeBooks folder for details.
The CodeBooks folder is a top level folder of the \code{PDS} package.  To see where the \code{PDS} package is
installed on your machine, type the following at the \code{R} prompt: \code{system.file(package = "PDS")}
}
\examples{
dim(marscrater_pds)
}
\keyword{datasets}

