% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{NESARC}
\alias{NESARC}
\title{National Epidemiologic Survey on Alcohol and Related Conditions}
\format{A data frame with 43093 observations on 3008 variables}
\source{
\url{http://pubs.niaaa.nih.gov/publications/AA70/AA70.htm}
}
\usage{
NESARC
}
\description{
NESARC
}
\details{
See the code books in inst/CodeBooks for details
}
\examples{
dim(NESARC)
}
\keyword{datasets}

