% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{OOL}
\alias{OOL}
\title{Outlook On Life Survey}
\format{A data frame with 2294 observations on 436 variables}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/ICPSR/studies/35348?q=outlook+on+LIfe&amp;searchSource=find-analyze-home&amp;sortBy=}
}
\usage{
OOL
}
\description{
Outlook on Life Surveys, 2012 (ICPSR 35348)
}
\details{
See the code books in the CodeBooks folder for details.
The CodeBooks folder is a top level folder of the \code{PDS} package.
To see where the \code{PDS} package is installed on your machine, type
the following at the \code{R} prompt:
\code{system.file(package = "PDS")}
}
\examples{
dim(OOL)
}
\keyword{datasets}

