% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pds-package.R
\docType{package}
\name{pds-package}
\alias{pds}
\alias{pds-package}
\title{pds: Probability for Data Science}
\description{
Probability prerequisites for data science courses, written as Quarto fragments that host sites include through a git submodule.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Morrison-Lab/pds}
  \item Report bugs at \url{https://github.com/Morrison-Lab/pds/issues}
}

}
\author{
\strong{Maintainer}: Ezra Morrison \email{d.ezra.morrison@gmail.com}

Authors:
\itemize{
  \item Ezra Morrison \email{d.ezra.morrison@gmail.com}
}

}
\keyword{internal}
