% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{signdist}
\alias{signdist}
\title{Sign Distance}
\format{\preformatted{'data.frame':	30 obs. of  2 variables:
 $ Age     : int  18 20 22 23 23 25 27 28 29 32 ...
 $ Distance: int  510 590 560 510 460 490 560 510 460 410 ...
}}
\source{
\url{https://oli.cmu.edu/jcourse/workbook/activity/page?context=434b8b4980020ca60017028aafe3258a}
}
\usage{
signdist
}
\description{
For n=30 drivers, the driver's age and the maximum distance at which the driver can read a highway sign at night.
}
\examples{
dim(signdist)
}
\keyword{datasets}

