% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDS.R
\docType{data}
\name{marscrater_pds}
\alias{marscrater_pds}
\title{Mars Craters}
\format{A data frame with 384343 observations on 10 variables}
\usage{
marscrater_pds
}
\description{
marscrater
}
\details{
See the code books in inst/CodeBooks for details
}
\examples{
dim(marscrater_pds)
}
\keyword{datasets}

